# storage/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "This is swagger documentation for store store",
    "version": "0.0.1",
    "title": "Store Backend",
    "termsOfService": "http://swagger.io/terms/",
    "contact": {
      "email": "miqdad.farcha@store.backend.fdc"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "securityDefinitions": {
    "bearerAuth": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  },
  "host": "store.backend.fdc",
  "basePath": "/",
...
  "schemes": [
    "http",
    "https"
  ],
  "paths": {
...
    "/prize-setting/add": {
      "post": {
        "tags": [
          "prize-setting"
        ],
        "summary": "Add New Prize Setting",
        "description": "",
        "operationId": "addPrizeSetting",
        "produces": [
          "application/json"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "name",
            "in": "formData",
            "description": "The name ",
            "required": true,
            "type": "string"
          },
          {
            "name": "description",
            "in": "formData",
            "description": "The description.",
            "required": true,
            "type": "string"
          },
          {
            "name": "prizes",
            "in": "formData",
            "description": "The prizes.",
            "required": true,
            "type": "integer"
          },
          {
            "name": "type",
            "in": "formData",
            "description": "The type.",
            "required": false,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "headers": {
              "X-Expires-After": {
                "type": "string",
                "format": "date-time",
                "description": "date in UTC when token expires"
              },
              "X-Rate-Limit": {
                "type": "integer",
                "format": "int32",
                "description": "calls per hour allowed by the user"
              }
            },
            "schema": {
              "type": "string"
            }
          },
          "400": {
            "description": "Invalid data supplied"
          },
          "401": {
            "description": "Unauthorized access"
          }
        }
      }
    },
    "/prize-setting/{id}": {
      "put": {
        "tags": [
          "prize-setting"
        ],
        "summary": "Update Prize Setting Record in the system",
        "description": "",
        "operationId": "updatePrizeSetting",
        "produces": [
          "application/json"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The Prize Setting ID ",
            "required": true,
            "type": "integer"
          },
          {
            "name": "name",
            "in": "formData",
            "description": "The name ",
            "required": true,
            "type": "string"
          },
          {
            "name": "description",
            "in": "formData",
            "description": "The description.",
            "required": true,
            "type": "string"
          },
          {
            "name": "prizes",
            "in": "formData",
            "description": "The prizes.",
            "required": true,
            "type": "integer"
          },
          {
            "name": "type",
            "in": "formData",
            "description": "The type.",
            "required": false,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "headers": {
              "X-Expires-After": {
                "type": "string",
                "format": "date-time",
                "description": "date in UTC when token expires"
              },
              "X-Rate-Limit": {
                "type": "integer",
                "format": "int32",
                "description": "calls per hour allowed by the user"
              }
            },
            "schema": {
              "type": "string"
            }
          },
          "400": {
            "description": "Invalid data supplied"
          },
          "401": {
            "description": "Unauthorized access"
          }
        }
      }
    },
    "/prize-setting/list": {
      "get": {
        "tags": [
          "prize-setting"
        ],
        "summary": "get Prize Setting list in the system",
        "description": "",
        "operationId": "getAllPrizeSetting",
        "produces": [
          "application/json"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "The Prize Setting Page",
            "required": false,
            "type": "integer"
          },
          {
            "name": "pageSize",
            "in": "query",
            "description": "The Prize Setting Page Size ",
            "required": false,
            "type": "integer"
          }
        ],
        "responses": {
          "default": {
            "description": "successful operation"
          }
        }
      }
    },
    "/prize-settingg/{id}": {
      "get": {
        "tags": [
          "prize-setting"
        ],
        "summary": "Get a prize setting details",
        "description": "",
        "operationId": "getPrizeSetting",
        "produces": [
          "application/json"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The prize setting ID",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "headers": {
              "X-Expires-After": {
                "type": "string",
                "format": "date-time",
                "description": "date in UTC when token expires"
              },
              "X-Rate-Limit": {
                "type": "integer",
                "format": "int32",
                "description": "calls per hour allowed by the user"
              }
            },
            "schema": {
              "type": "string"
            }
          },
          "400": {
            "description": "Invalid data supplied"
          },
          "401": {
            "description": "Unauthorized access"
          }
        }
      }
    },
    "/prize-settings/{id}": {
      "delete": {
        "tags": [
          "prize-setting"
        ],
        "summary": "delete Prize Setting from the system",
        "description": "",
        "operationId": "deletePrizeSetting",
        "produces": [
          "application/json"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The user address ID",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "default": {
            "description": "successful operation"
          }
        }
      }
    },
...
    "/manufacturer/add": {
      "post": {
        "tags": [
          "manufacturer"
        ],
        "summary": "Add Manufacturer to the system",
        "description": "",
        "operationId": "addManufacturer",
        "produces": [
          "application/json"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "name",
            "in": "formData",
            "description": "The name ",
            "required": true,
            "type": "string"
          },
          {
            "name": "description",
            "in": "formData",
            "description": "The description.",
            "required": false,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "headers": {
              "X-Expires-After": {
                "type": "string",
                "format": "date-time",
                "description": "date in UTC when token expires"
              },
              "X-Rate-Limit": {
                "type": "integer",
                "format": "int32",
                "description": "calls per hour allowed by the user"
              }
            },
            "schema": {
              "type": "string"
            }
          },
          "400": {
            "description": "Invalid data supplied"
          },
          "401": {
            "description": "Unauthorized access"
          }
        }
      }
    },
...
    "/manufacturer/delete": {
      "delete": {
        "tags": [
          "manufacturer"
        ],
        "summary": "delete manufacturer from the system",
        "description": "",
        "operationId": "deleteManufacturer",
        "produces": [
          "application/json"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "manufacturer_id",
            "in": "formData",
            "description": "The manufacturer ID",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "default": {
            "description": "successful operation"
          }
        }
      }
    },
...
    "/user-spinwheel/view-by-user-id": {
      "get": {
        "tags": [
          "user-spinwheel"
        ],
        "summary": "Get user spinwheel details",
        "description": "",
        "operationId": "getPrizeSetting",
        "produces": [
          "application/json"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "query",
            "description": "The user ID",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "headers": {
              "X-Expires-After": {
                "type": "string",
                "format": "date-time",
                "description": "date in UTC when token expires"
              },
              "X-Rate-Limit": {
                "type": "integer",
                "format": "int32",
                "description": "calls per hour allowed by the user"
              }
            },
            "schema": {
              "type": "string"
            }
          },
          "400": {
            "description": "Invalid data supplied"
          },
          "401": {
            "description": "Unauthorized access"
          }
        }
      }
    },
...
  },
...
}